gradio-app/requirements.txt
Flask
tensorflow==2.12.0
numpy
Pillow
gradio==3.5.0
httpx==0.23.0
httpcore==0.15.0